+++ src/components/TrialDetail/DetailBoard/TrialInfo/index.tsx
@@ -1,6 +1,7 @@
 import ProgressBar from "./ProgressBar";
 import TrialContent from "./TrialContent";
-import type { Trial } from "@/features/trials/type";
+import type { Trial } from "@/components/types/Trial";
+import Challenges from "./Challenges";
 interface acceptProps {
   trial: Trial;
 }
@@ -13,22 +14,25 @@ export default function TrialInfo(props: acceptProps) {
   }
 
   return (
-    <div className="bg-card-bg rounded-xl p-5 flex flex-col justify-between gap-4 max-lg:flex-col-reverse ">
-      <div className="flex w-full justify-between max-lg:max-w-none max-lg:flex-col-reverse ">
-        <div className="flex gap-4 w-fit max-lg:justify-start">
-          {trial.isPublic && (
-            <p className="text-label rounded-full bg-gray-200 font-semibold text-black w-fit h-fit py-0.5 px-2.5">
-              公開
+    <div className="flex flex-col gap-6">
+      <div className="bg-card-bg rounded-xl p-5 flex flex-col justify-between gap-4 max-lg:flex-col-reverse ">
+        <div className="flex w-full justify-between max-lg:max-w-none max-lg:flex-col-reverse ">
+          <div className="flex gap-4 w-fit max-lg:justify-start">
+            {trial.isPublic && (
+              <p className="text-label rounded-full bg-gray-200 font-semibold text-black w-fit h-fit py-0.5 px-2.5">
+                公開
+              </p>
+            )}
+            <p className="text-xs rounded-full bg-gray-200 font-semibold text-black w-fit h-fit py-0.5 px-2.5">
+              {trial.trialState}
             </p>
-          )}
-          <p className="text-xs rounded-full bg-gray-200 font-semibold text-black w-fit h-fit py-0.5 px-2.5">
-            {trial.trialState}
-          </p>
+          </div>
+          <ProgressBar trial={trial} />
         </div>
-        <ProgressBar trial={trial} />
+        {/* right */}
+        <TrialContent trial={trial} />
       </div>
-      {/* right */}
-      <TrialContent trial={trial} />
+      <Challenges challenges={trial.challenges} />
     </div>
   );
 }
